requirements.txt
certifi==2025.1.31
polygon-api-client==1.12.3
python-dotenv==1.0.0
urllib3==2.3.0
websockets==14.2
